Move `{Free,}RegionRelations` and `FreeRegionMap` out of `rustc_middle`
